+++ main.tf
@@ -5,6 +5,7 @@ resource "azurerm_virtual_network" "vnet" {
   tags                = var.tags
   address_space       = var.address_space
   dns_servers         = var.dns_servers
+  bgp_community       = var.bgp_community
 
   dynamic "ddos_protection_plan" {
     for_each = var.ddos_protection_plan_id != null ? [""] : []