+++ variables.tf
@@ -24,6 +24,11 @@ variable "ddos_protection_plan_id" {
   default     = null
   description = "The ID of DDoS Protection Plan."
 }
+variable "bgp_community" {
+  type        = string
+  default     = null
+  description = "The BGP community attribute in format <as-number>:<community-value>. The as-number segment is the Microsoft ASN, which is always 12076 for now."
+}
 variable "subnets" {
   type        = any
   default     = []